St.6 Refactor code Add REST API endpoints description Rename controllers (new name based on his role)
Fix some bugs.

+++ src/main/java/ru/javaops/topjava2/repository/VoteRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface VoteRepository extends JpaRepository<Vote, Integer> {
+public interface VoteRepository extends JpaRepository<Vote, Long> {
     Optional<Vote> findByUserIdAndDate(Integer userId, LocalDate date);
 
     List<Vote> findByDate(LocalDate date);
